requirements.txt
streamlit>=1.29.0
psycopg2-binary>=2.9.9
pymongo>=4.6.0
pandas>=2.1.4
faker>=20.1.0

